Move database to native/db
